# swagger.yaml
openapi: "3.0.2"
info:
  title: "Puppy Shop"
  version: "1.0"
servers:
  - url: http://localhost:10778/
    description: shop server
paths:
  /users/:
    post:
      description: register and validate new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dryRun:
                  type: boolean
                  example: true
                israeliID:
                  type: integer
                  example: 123456789
                email:
                  type: string
                  example: user@gmail.com
                password:
                  type: string
                  example: q123
                first_name:
                  type: string
                  example: first-name
                last_name:
                  type: string
                  example: last-name
                city:
                  type: string
                  example: Ashdod
                street:
                  type: string
                  example: Katz
              required:
                - dryRun
                - israeliID
                - email
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: false
                  exists:
                    type: boolean
                    example: false
        "201":
          description: created
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: false
                  token:
                    type: string
                    example: token
                  refreshToken:
                    type: string
                    example: token
        "400":
          description: Id and/or email already exist in our data
        "406":
          description: missing some info
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/cities:  
    get:
      description: get 10 big cities (registration and order form)
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: false
                  cities:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 2
                        name:
                          type: string
                          example: Ashdod

        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/login:
    post:
      description: "login"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@gmail.com
                password:
                  type: string
                  example: q123
              required:
                - email
                - password
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: false
                  token:
                    type: string
                    example: token
                  refreshToken:
                    type: string
                    example: token
        "406":
          description: missing some info
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: true
                  msg:
                    type: string
                    example: missing some info
        "404":
          description: not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: true
                  msg:
                    type: string
                    example: user not found
        "401":
          description: unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: true
                  msg:
                    type: string
                    example: wrong password
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/logout:
    put:
      description: logout
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          $ref: "#/components/responses/user"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/check:
    get:
      description: validate token
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          $ref: "#/components/responses/user"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    token:
      type: apiKey
      in: header
      name: token
    refreshToken:
      type: apiKey
      in: header
      name: refreshToken
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: true
              msg:
                type: string
                example: error
...
    user:
      description: user
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: false
              user:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  israeliID:
                    type: integer
                    example: 123456789
                  first_name:
                    type: string
                    example: sharon
                  last-name:
                    type: string
                    example: brown  
                  email:
                    type: string
                    example: user@gmail.com
                  role:
                    type: integer
                    example: 2  
                  city:
                    type: string
                    example: Ashdod
                  street:
                    type: string
                    example: Katz
                  isLogin:
                    type: boolean  
                    example: false
...